artstractaguiar/requirements.txt
Flask==3.0.3
Flask_Login==0.6.3
numpy==1.26.4
Pillow==10.3.0
Werkzeug==3.0.3
